added UI comp, working on lobby

+++ Test/UIObjectFactory.h
@@ -33,7 +33,7 @@ public:
 
 	/// Creates a new UI Object
 	UIObject * Spawn(UIType PID);
-	UIObject * Spawn(UIType PID, float x, float y);
+	UIObject * Spawn(UIType PID, int x, int y);
 
 };
